Move mock definition to its own file
